RST-3494 Combining et1a pdf with et1 for multiples
