+++ .rubocop.yml
@@ -2,7 +2,7 @@ inherit_from:
   - .rubocop_todo.yml
 
 AllCops:
-  TargetRubyVersion: 2.3
+  TargetRubyVersion: 2.7
   # Exclude anything that isn't really part of our code.
   # rails_helper is excluded because it's full of solecisms, but it's mostly
   # generated code and copy-and-pasted snipets from READMEs.
@@ -24,9 +24,6 @@ AllCops:
     - 'Gemfile'
     - 'config/spring.rb'
 
-
-Rails:
-  Enabled: true
 
 # Allow long lines in specs, as it's almost impossible to fit RSpec's
 # expectations into 80 characters.
@@ -57,7 +54,7 @@ Style/AsciiComments:
   Enabled: true
   Exclude:
     - 'spec/**/*'
-Layout/TrailingBlankLines:
+Layout/TrailingEmptyLines:
   Enabled: true
 # Don't worry about long methods in specs.
 Metrics/MethodLength:
@@ -83,7 +80,7 @@ Layout/AccessModifierIndentation:
 
 # Just indent parameters by two spaces. It's less volatile if methods change,
 # and there's less busy work lining things up.
-Layout/AlignParameters:
+Layout/ParameterAlignment:
   EnforcedStyle: with_fixed_indentation
 
 Style/ClassAndModuleChildren:
