+++ Gemfile
@@ -80,7 +80,7 @@ group :test do
   gem 'faker', '~> 2.15'
   gem 'webmock', '~> 3.11'
   gem 'et_fake_acas_server', '~> 0.1'
-  gem 'gov_fake_notify', git: 'https://github.com/hmcts/gov_fake_notify.git', tag: 'v1.1.1'
+  gem 'gov_fake_notify', git: 'https://github.com/hmcts/gov_fake_notify.git', tag: 'v1.1.2'
 end
 
 # Windows does not include zoneinfo files, so bundle the tzinfo-data gem
