+++ app/models/uploaded_file.rb
@@ -8,9 +8,9 @@ class UploadedFile < ApplicationRecord
 
   has_many :response_uploaded_files, dependent: :destroy
 
-  scope :not_hidden, -> { where('filename NOT LIKE ? AND filename NOT LIKE ? AND filename NOT LIKE ? AND filename NOT ILIKE ?', 'acas_%', 'et1_%.txt', 'et1a_%.txt', 'et1_%.rtf') }
+  scope :not_hidden, -> { where('filename NOT LIKE ? AND filename NOT LIKE ? AND filename NOT LIKE ? AND filename NOT ILIKE ? AND filename NOT ILIKE ?', 'acas_%', 'et1_%.txt', 'et1a_%.txt', 'et1_%.rtf', 'et1_%_trimmed.pdf') }
   scope :not_pdf, -> { where('filename NOT LIKE ?', 'et1_%.pdf') }
-  scope :et1_pdf, -> { where('filename LIKE ?', 'et1_%.pdf') }
+  scope :et1_pdf, -> { where('filename LIKE ? AND filename NOT LIKE ?', 'et1_%.pdf', 'et1_%_trimmed.pdf') }
   scope :et1_rtf, -> { not_hidden.where('filename ILIKE ?', '%.rtf') }
   scope :et1_csv, -> { where('filename LIKE ?', 'et1a_%.csv') }
   scope :et3_pdf, -> { where('filename LIKE ?', 'et3_atos_export.pdf') }
