+++ spec/factories/claim_factory.rb
@@ -21,7 +21,7 @@ FactoryBot.define do
     jurisdiction { 2 }
     office_code { 22 }
     date_of_receipt { Time.zone.now }
-    pdf_template_reference { "et1-v2-en" }
+    pdf_template_reference { "et1-v3-en" }
     email_template_reference { "et1-v1-en" }
     claim_details { 'claim details field' }
     miscellaneous_information { 'miscellaneous information' }
