+++ spec/factories/json/json_claim_factory.rb
@@ -21,7 +21,7 @@ FactoryBot.define do
       secondary_claimant_traits { [:mr_first_last] }
       secondary_respondent_traits { [:full] }
       primary_representative_traits { [:full] }
-      pdf_template { 'et1-v2-en' }
+      pdf_template { 'et1-v3-en' }
       email_template { 'et1-v1-en' }
     end
 
@@ -70,7 +70,7 @@ FactoryBot.define do
     end
 
     trait :with_welsh_pdf do
-      pdf_template { 'et1-v2-cy' }
+      pdf_template { 'et1-v3-cy' }
     end
 
     trait :with_welsh_email do
@@ -146,7 +146,7 @@ FactoryBot.define do
       send_claim_to_whistleblowing_entity { false }
       miscellaneous_information { 'Miscellaneous Information' }
       is_unfair_dismissal { false }
-      pdf_template_reference { "et1-v2-en" }
+      pdf_template_reference { "et1-v3-en" }
       email_template_reference { "et1-v1-en" }
       confirmation_email_recipients { ['confirmation_recipient@digital.justice.gov.uk'] }
     end
