+++ spec/support/landing_folder_support/staging_folder.rb
@@ -68,7 +68,7 @@ module EtApi
         EtApi::Test::FileObjects::Et3PdfFile.new extract_to_tempfile(filename), template: template, lookup_root: 'response_pdf_fields'
       end
 
-      def et1_pdf_file(filename, template: 'et1-v2-en')
+      def et1_pdf_file(filename, template: 'et1-v3-en')
         EtApi::Test::FileObjects::Et1PdfFile.new extract_to_tempfile(filename), template: template, lookup_root: 'claim_pdf_fields'
       end
